Move reporter to examples
