--- src/telliot/reporter/__init__.py
@@ -1,2 +1,0 @@
-"""Subpackage for TellorDAO-supported Reporters used by
-the `telliot` CLI."""
