+++ src/telliot_feed_examples/reporters/profitcalc.py
