+++ tests/test_profit_calc.py
@@ -2,7 +2,7 @@
 Test covering Pytelliot EVM contract connection utils.
 """
 import pytest
-from telliot.reporter.profitcalc import profitable
+from telliot_feed_examples.reporters.profitcalc import profitable
 
 
 @pytest.fixture
